moving the map function into a separate component
