+++ src/components/paymentTable/paymentTable.js
@@ -1,5 +1,9 @@
 import React from 'react';
 import './paymentTable.scss';
+import '../../assets/main.css';
+import { useTranslation } from 'react-i18next';
+import booksRender from './booksRender';
+
 import {
   MDBContainer,
   MDBRow,
@@ -10,68 +14,35 @@ import {
 } from 'mdbreact';
 
 export default function PaymentTable() {
-  const books = [
-    {
-      title: 'book1',
-      copies: 1,
-      price: '$',
-    },
-    {
-      title: 'book2',
-      copies: 1,
-      price: '$',
-    },
-    {
-      title: 'book3',
-      copies: 1,
-      price: '$',
-    },
-  ];
-  const paymentTable = {
-    booksOrder: 'Your Books Order',
-    books: 'Books',
-    copies: 'Copies',
-    price: 'Price',
-    total: 'Total',
-  };
-
+  const { t } = useTranslation();
   return (
     <MDBContainer>
       <MDBRow>
-        <MDBCol md="4">
+        <MDBCol md="12">
           <div className="text-center">
-            <p className="h5 mt-5 gray-text">{paymentTable.booksOrder}</p>
-            <hr className="border-4 my-3 mx-5 bg-primary"></hr>
+            <p className="h5 mt-5 text-gray-800">
+              {t('PaymentPage.booksOrder')}
+            </p>
+            <hr className="border-4 my-3 mx-4 bg-primary"></hr>
           </div>
           <MDBTable>
             <MDBTableHead>
               <tr>
                 <th></th>
-                <th>{paymentTable.books}</th>
-                <th>{paymentTable.copies}</th>
-                <th>{paymentTable.price}</th>
+                <th>{t('PaymentPage.books')}</th>
+                <th>{t('PaymentPage.copies')}</th>
+                <th>{t('PaymentPage.price')}</th>
               </tr>
             </MDBTableHead>
             <MDBTableBody>
-              {books.map((book) => {
-                return (
-                  <tr key="index">
-                    <td>
-                      <input type="checkbox" />
-                    </td>
-                    <td>{book.title}</td>
-                    <td>{book.copies}</td>
-                    <td>{book.price}</td>
-                  </tr>
-                );
-              })}
+              {booksRender()}
               <tr>
                 <td></td>
                 <td>
-                  <p className="h6">{paymentTable.total}</p>
+                  <p className="h6">{t('PaymentPage.total')}</p>
                 </td>
-                <td>{books.copies}</td>
-                <td>{books.price}</td>
+                <td>{}</td>
+                <td>{}</td>
               </tr>
             </MDBTableBody>
           </MDBTable>
